pyproject.toml
[tool.poetry]
name = "odkbot"
version = "0.1.0"
description = ""
authors = ["Carlo De Pieri <depieri.carlo@gmail.com>"]
license = "GPT-3"

packages = [
  { include = "odkbot" }
]

[tool.poetry.scripts]
odkbot_prod = 'odkbot.main:run_prod'
odkbot_dev = 'odkbot.main:run_dev'

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.3"

[tool.poetry.dev-dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
